fix: omit external tools from server builds
