+++ .changeset/strip-external-tools.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/x-generative-compiler": patch
+---
+
+fix: omit externalTool entries from server toolkits
