+++ apps/docs/content/docs/(reference)/api-reference/utilities/miscellaneous.mdx
@@ -54,8 +54,9 @@ Marks a generative toolkit entry as an externally executed backend tool.
 
 Use this when another system (for example a backend route or LangGraph node)
 already defines and executes the tool, but assistant-ui should render its
-tool calls. The use-generative compiler converts `execute: externalTool()`
-into `type: "backend"` and strips schema/executor metadata from both builds.
+tool calls. The use-generative compiler omits `execute: externalTool()`
+entries from the server build and keeps a `type: "backend"` renderer on the
+client build.
 
 ```ts
 function externalTool(): never;
