+++ examples/with-artifacts/app/page.tsx
@@ -4,7 +4,6 @@ import { Thread } from "@/components/assistant-ui/thread";
 import {
   AssistantRuntimeProvider,
   Tools,
-  type Toolkit,
   useAui,
   useAuiState,
   AuiProvider,
@@ -12,30 +11,9 @@ import {
 } from "@assistant-ui/react";
 import { useChatRuntime } from "@assistant-ui/react-ai-sdk";
 import type { ToolCallMessagePart } from "@assistant-ui/react";
-import { TerminalIcon, CodeIcon, EyeIcon } from "lucide-react";
+import { CodeIcon, EyeIcon } from "lucide-react";
 import { useState } from "react";
-import { z } from "zod";
-
-const toolkit = {
-  render_html: {
-    description:
-      "Whenever the user asks for HTML code, call this function. The user will see the HTML code rendered in their browser.",
-    parameters: z.object({
-      code: z.string(),
-    }),
-    execute: async () => {
-      return {};
-    },
-    render: () => {
-      return (
-        <div className="bg-primary text-primary-foreground my-2 inline-flex items-center gap-2 rounded-full border px-4 py-2">
-          <TerminalIcon className="size-4" />
-          render_html(&#123; code: &quot;...&quot; &#125;)
-        </div>
-      );
-    },
-  },
-} satisfies Toolkit;
+import toolkit from "./toolkit";
 
 function ArtifactsView() {
   const [tab, setTab] = useState<"source" | "preview">("source");
